src/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.11
django-filter==24.2
django-parler==2.3
django-rosetta==0.9.8
djangorestframework==3.15.1
idna==3.7
importlib_metadata==7.1.0
Markdown==3.6
pillow==10.3.0
polib==1.2.0
requests==2.32.3
sqlparse==0.5.0
stripe==9.12.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
zipp==3.19.0
